Add base for the admin area for context.
In context's tests, we fake the current_user object.
At the moment, we assume there's at least a current_user object set by
the host application and that it has a context_admin boolean field set
on it. This is how we can determine if a user is a context admin and
redirect them away accordingly if they're not.